+++ src/analysis/old/analyze_gearbox_origin_all_v2_overflow.rs
@@ -794,7 +794,7 @@ impl<'a> GearboxSim<'a> {
         let total_rows = input_vec.rows();
         // print every 1% or every 60s
         let mut next_print_percent = total_rows / 100;
-        let mut next_print_time = TIME_TO_LOG as u64;
+        let mut next_print_time = TIME_TO_LOG;
         //each data size if 8 bytes and there are 512 rows in a subarray
         let mut overflow_count_8_512 = 0;
         let mut overflow_count_8_256 = 0;
@@ -810,7 +810,7 @@ impl<'a> GearboxSim<'a> {
                 let speed = target_id as f32 / min;
                 tracing::trace!("{target_id} of {total_rows} rows processed, time eclips: {min:.2}, estimate remaining time:{min_r:.2},speed: {speed} rows per min");
                 next_print_percent = target_id + total_rows / 100;
-                next_print_time = now.elapsed().as_secs() + TIME_TO_LOG as u64;
+                next_print_time = now.elapsed().as_secs() + TIME_TO_LOG;
                 if stop_signal::read() {
                     break;
                 }
